Refactored everything in subfolders
