+++ frontend/src/app/app.routes.ts
@@ -1,12 +1,12 @@
 import { Routes } from '@angular/router';
-import { AccountComponent } from './account.component';
-import { HomeComponent } from './home.component';
-import { ShopComponent } from './shop.component';
-import { CartComponent } from './cart.component';
-import { AdminComponent } from './admin.component';
-import { LoginComponent } from './login.component';
-import { RegisterComponent } from './register.component';
-import { ProductDetailComponent } from './product-detail.component';
+import { AccountComponent } from './pages/account/account.component';
+import { HomeComponent } from './pages/home/home.component';
+import { ShopComponent } from './pages/shop/shop.component';
+import { CartComponent } from './pages/cart/cart.component';
+import { AdminComponent } from './pages/admin/admin.component';
+import { LoginComponent } from './pages/login/login.component';
+import { RegisterComponent } from './pages/register/register.component';
+import { ProductDetailComponent } from './pages/product-detail/product-detail.component';
 import { authGuard } from './guards/auth.guard';
 import { adminGuard } from './guards/admin.guard';
