+++ frontend/src/app/components/menu-card/menu-card.component.html
