+++ frontend/src/app/components/navbar/navbar.component.scss
