+++ frontend/src/app/components/product-card/product-card.component.scss
