+++ frontend/src/app/pages/admin/admin.component.scss
