+++ frontend/src/app/pages/cart/cart.component.scss
