+++ frontend/src/app/pages/register/register.component.scss
