+++ frontend/src/app/pages/shop/shop.component.scss
